backend/requirements.txt
asgiref==3.7.2
distlib==0.3.6
Django==4.2.3
djangorestframework==3.14.0
filelock==3.12.2
platformdirs==3.8.1
pytz==2023.3
sqlparse==0.4.4
virtualenv==20.23.1
django-cors-headers==4.2.0
mysqlclient==2.2.0
pillow==10.4.0
requests==2.31.0
